+++ orchestrator-bundle/orc8r-ha-operator/requirements.txt
@@ -1,4 +1,3 @@
 git+https://github.com/canonical/operator/#egg=ops
-ops-lib-pgsql
 lightkube
 lightkube-models
